1-print_binary.c

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -6,7 +6,7 @@
  * @b: An unsigned int b
  * Return: power
  */
-int _pow(int a, int b);
+long int _pow(int a, int b);
 
 /**
  * print_binary - the function that convert binary to int
@@ -15,8 +15,8 @@ int _pow(int a, int b);
  */
 void print_binary(unsigned long int n)
 {
-	int ne = (int)n;
-	int i = 0, power = 1, comp = 0;
+	long int ne = (long int)n;
+	long int i = 0, power = 1, comp = 0;
 
 	if (n == 0)
 	{
@@ -72,9 +72,9 @@ void print_binary(unsigned long int n)
  * @b: An unsigned int b
  * Return: power
  */
-int _pow(int a, int b)
+long int _pow(int a, int b)
 {
-	int result = 1;
+	long int result = 1;
 
 	while (b--)
 		result *= a;
